Organizing keymap items for cleaner edits and copying. Added boost functionality to rapid fire.
