+++ keyboards/recorev3/rules.mk
@@ -1,6 +1,5 @@
 # Debounce Configuration
-# DEBOUNCE_TYPE = custom # asym, instant, then defer
-SRC += debounce.c
+SRC += asym_instant_defer_pk.c
 
 # RGB Configuration
 WS2812_DRIVER = vendor
